refactoring + edit-point-view
